+++ Delete.php
@@ -14,13 +14,13 @@ declare(strict_types=1);
 
 namespace Hector\Query;
 
-use Hector\Query\Component\Order;
+use Hector\Connection\Bind\BindParamList;
 use Hector\Query\Clause\BindParams;
 use Hector\Query\Clause\From;
-use Hector\Query\Clause\Where;
 use Hector\Query\Clause\Limit;
+use Hector\Query\Clause\Where;
 use Hector\Query\Component\EncapsulateHelperTrait;
-use Hector\Connection\Bind\BindParamList;
+use Hector\Query\Component\Order;
 
 class Delete implements StatementInterface
 {