+++ Sort/SortInterface.php
@@ -1,0 +1,25 @@
+<?php
+/*
+ * This file is part of Hector ORM.
+ *
+ * @license   https://opensource.org/licenses/MIT MIT License
+ * @copyright 2026 Ronan GIRON
+ * @author    Ronan GIRON <https://github.com/ElGigi>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code, to the root.
+ */
+
+declare(strict_types=1);
+
+namespace Hector\Query\Sort;
+
+use Hector\Query\QueryBuilder;
+
+interface SortInterface
+{
+    /**
+     * Apply sort to query builder.
+     */
+    public function apply(QueryBuilder $builder): void;
+}