+++ Statement/Conditions.php
@@ -14,9 +14,9 @@ declare(strict_types=1);
 
 namespace Hector\Query\Statement;
 
-use Hector\Query\Clause\Where;
-use Hector\Query\Clause\Having;
 use Hector\Connection\Bind\BindParamList;
+use Hector\Query\Clause\Having;
+use Hector\Query\Clause\Where;
 use Hector\Query\StatementInterface;
 
 class Conditions implements StatementInterface